エクスポート経路の変更
