+++ src/app/memo/_components/index.tsx
@@ -1,0 +1,1 @@
+export * from './RichEditor'
