+++ src/components/template/Layout/Header/Header.tsx
@@ -1,0 +1,10 @@
+import { Group } from '@/components/ui'
+import { Logo } from './Logo'
+
+export const Header = () => {
+  return (
+    <Group h="100%" px="md">
+      <Logo />
+    </Group>
+  )
+}
