+++ src/components/template/Layout/Header/Logo.tsx
@@ -1,0 +1,11 @@
+import { Anchor, BaseTitle } from '@/components/ui'
+import { meta } from '@/config'
+import Link from 'next/link'
+
+export const Logo = () => {
+  return (
+    <Anchor c="dark" component={Link} href="/" underline="never">
+      <BaseTitle>{meta.title.default}</BaseTitle>
+    </Anchor>
+  )
+}
