+++ src/components/template/Layout/Header/index.tsx
@@ -1,0 +1,1 @@
+export { Header } from './Header'
