+++ src/components/template/Layout/index.tsx
@@ -1,0 +1,1 @@
+export { Layout } from './Layout'
